+++ src/share/data/coordinate/relative-coordinate/RelativeCoordinate.ts
@@ -11,6 +11,7 @@ import { Settings } from "@share/config/Settings";
  * The x, y and z coordinates must be within 0 and the chunk size minus 1.
  * @throws {RangeError} If the x, y or z coordinate is not within 0 and the chunk size minus 1.
  * @throws {TypeError} If decimals are used as an input.
+ * @deprecated Use Vector3 instead
  */
 
 export class RelativeCoordinate implements IRelativeCoordinate {